Remove support for Python 2.
This includes adding Python 3 features, like f-strings, that aren't
technically necessary but that were only excluded before for Python 2
support.

This is backwards-incompatible, so it comes with a major version bump.

Given the length of time this project has been live with no interface
changes, it makes sense to go to 1.0.0 rather than staying at 0.x.y.