Run terratest for all Python versions.
This includes an addition to terratest that logs which version
the 'python' binary points to. Since that's the binary run by
the external data resource in the example, it tells us what
version of Python is running the decorator.

To get tox to run both pylint and terratest when the 'py' env
was set (the pipeline uses this to manage python versions inside
its build servers), both commands had to be put in the default
testenv. Experimentation showed that the generative syntax could
create more testenvs for terratest, but it wouldn't select them
from the current version of Python when the 'py' env was set. It
only does that for the default testenv. See azure_pipelines.yaml
for details on how 'py' is used. This requires a slightly funky
--rcfile flag for pylint, but otherwise works fine.